updates
frustrating as it seems to work with one test and then doesn't with the next. SignUp, signIn, signOut and status change is in there. External providers login is not yet set-up.

Very grateful if anyone sees why it seems working one minute (network gave me 200 for everything), then i do it again and it just doesn't even go into the function). I get that the 'else' part in signedIn function can be left out, but it didn't work even if I left it out.